rename gallery components
