+++ src/components/FilterNav.jsx
@@ -2,10 +2,10 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import Nav from 'react-bootstrap/Nav';
 import { Image, CloudinaryContext, Transformation } from 'cloudinary-react';
-import { filters } from '../utils/sorting';
+import { filters } from '../utils/filters';
 import '../css/galleryNav.scss';
 
-class GalleryNav extends Component {
+class FilterNav extends Component {
 
   constructor (props) {
     super(props);
@@ -17,10 +17,10 @@ class GalleryNav extends Component {
   }
 
   componentDidMount () {
-    this.handleSortSelect(0);
+    this.handleFilterSelect(0);
   }
 
-  handleSortSelect = eventKey => {
+  handleFilterSelect = eventKey => {
     // console.log('Sort selected: ' + eventKey);
     this.setState({
       filterIndex: eventKey,
@@ -45,7 +45,12 @@ class GalleryNav extends Component {
     const activeFilter = filters[filterIndex];
     return (
       <CloudinaryContext cloudName="cantimaginewhy">
-        <Nav variant="pills" defaultActiveKey={0} className="gallery-nav justify-content-center" onSelect={this.handleSortSelect}>
+        <Nav 
+          variant="pills" 
+          defaultActiveKey={0} 
+          className="gallery-nav justify-content-center" 
+          onSelect={this.handleFilterSelect}
+        >
           <span className="label">Filter By:</span>
           {filters.map(filter => (
             <Nav.Item key={filter.name}>
@@ -87,9 +92,9 @@ class GalleryNav extends Component {
   }
 }
 
-GalleryNav.propTypes = {
+FilterNav.propTypes = {
   handleNavChange: PropTypes.func.isRequired,
   handleClearGallery: PropTypes.func
 };
 
-export default GalleryNav;
+export default FilterNav;
